moving login from window to modal - partially done, working out minor bugs
