+++ src/App.js
@@ -1,9 +1,9 @@
 //import modules/css
 import "./styles/App.css";
-import { Routes, Route, useNavigate } from "react-router-dom";
-import { useState } from "react";
-import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import { Routes, Route } from "react-router-dom";
+import { useState } from "react";
+import { ToastContainer } from "react-toastify";
 import { GlobalContextProvider } from "./Context/GlobalContext";
 
 //component imports
@@ -11,109 +11,34 @@ import HomePage from "./pages/Homepage";
 import RecordLibrary from "./pages/RecordLibrary";
 import Theme from "./Context/ThemeContext";
 import Layout from "./Layout";
-import FormRoute from "./../src/components/common/Form/FormRoute.jsx";
+import Modal from "./components/common/Modal.jsx";
 
-//firebase
-// import { app } from "./firebase/firebase-config.js";
-import {
-  getAuth,
-  signInWithEmailAndPassword,
-  createUserWithEmailAndPassword,
-} from "firebase/auth";
+
 
 export default function App() {
-  const [loginValues, setLogin] = useState({
-    email: "",
-    password: "",
-    showPassword: false,
-  });
 
-  const navigate = useNavigate();
-
-  const handleClickSubmitLogin = (id) => {
-    console.log("ID of form (login/signup):", id);
-    console.log("email:", loginValues.email);
-    console.log("Password:", loginValues.password);
-    const authentication = getAuth();
-
-    if (id === "register") {
-      createUserWithEmailAndPassword(
-        authentication,
-        loginValues.email,
-        loginValues.password
-      )
-        .then((response) => {
-          navigate("/");
-          sessionStorage.setItem(
-            "Auth Token",
-            response._tokenResponse.refreshToken
-          );
-        })
-        .catch((error) => {
-          console.log(error);
-          console.log(error.code);
-
-          if (error.code === "auth/email-already-in-use") {
-            toast.error("Email Already in Use");
-          }
-        });
-    }
-    if (id === "login") {
-      signInWithEmailAndPassword(
-        authentication,
-        loginValues.email,
-        loginValues.password
-      )
-        .then((response) => {
-          navigate("/");
-          sessionStorage.setItem(
-            "Auth Token",
-            response._tokenResponse.refreshToken
-          );
-        })
-        .catch((error) => {
-          console.log("error", error);
-          console.log("error.code", error.code);
-
-          if (error.code === "auth/wrong-password") {
-            toast.error(
-              "Password may have been incorrect, please check and try again"
-            );
-          }
-          if (error.code === "auth/user-not-found") {
-            toast.error("Invalid Email, try again");
-          }
-          if (error.code === "auth/too-many-requests") {
-            toast.error(
-              "Too Many requests. Access to this account has been temporarily disabled due to many failed login attempts. You can immediately restore it by resetting your password or you can try again later. "
-            );
-          }
-        });
-    }
-  };
+  //------ MODAL ------//
+  // By default, modal is an empty object {}
+  // When using setModal, the syntax is setModal({ modalName, modalData })
+  // modalName: Required; a string, decides which modal content to render
+  // modalData: Optional; an object, should contain some data you need for the modal
+  // Example 1 - setModal: line 14 @ './QuestionsAndAnswers/QAListEntry.jsx'
+  // Example 2 - redernModal: line 8 @ './helper/Modals/Modals.jsx'
+  const [modal, setModal] = useState({});
 
   return (
     <div className="App">
       <Theme>
         <GlobalContextProvider>
-          <Layout>
+          <Layout setModal={setModal}>
             <Routes>
               <Route path="/" element={<HomePage />} />
               <Route path="/records" element={<RecordLibrary />} />
-              <Route
-                path="/form/*"
-                element={
-                  <FormRoute
-                    handleClickSubmitLogin={handleClickSubmitLogin}
-                    loginValues={loginValues}
-                    setLogin={setLogin}
-                  />
-                }
-              />
             </Routes>
           </Layout>
         </GlobalContextProvider>
       </Theme>
+      <Modal modal={modal} />
       <ToastContainer toastStyle={{ backgroundColor: "black" }} />
     </div>
   );
