+++ src/pages/Homepage.js
@@ -1,13 +1,12 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 
 import React, { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import { toast } from "react-toastify";
+// import { useNavigate } from "react-router-dom";
 
 import useGlobalContext from "../Context/GlobalContext";
 
 export default function HomePage() {
-  let navigate = useNavigate();
+  // let navigate = useNavigate();
 
   const { exampleState, setExampleState } = useGlobalContext();
 
@@ -17,24 +16,7 @@ export default function HomePage() {
     setExampleState("setExampleState inside Homepage.js");
   }, []);
 
-  useEffect(() => {
-    let authToken = sessionStorage.getItem("Auth Token");
-    if (authToken) {
-      toast.success("Logged IN", {
-        hideProgressBar: false,
-        position: "bottom-left",
-      });
-      navigate("/");
-    }
-    //DON"T REMOVE: this can be useful to navigate to pages with dummy data upon login IF no user is detected
-    if (!authToken) {
-      toast.error("Not Logged in: Please Login to use all Features", {
-        hideProgressBar: true,
-        position: "top-left",
-      });
-      // navigate("/login");
-    }
-  }, []);
+
 
   return <h1>HomePage</h1>;
 }
